ANDKR-372 | 고메 검색 결과 관련 연동중 임시 저장

+++ dailyHOTEL/app/src/main/java/com/twoheart/dailyhotel/model/Gourmet.java
@@ -19,7 +19,7 @@ public class Gourmet extends Place
     public String category;
     public int categoryCode;
     public int categorySequence;
-    public float distance;
+    public double distance;
 
     private GourmetFilters mGourmetFilters;
